Proyecto Recetas FIN

+++ MF0968/recetas/src/main/java/com/ipartek/formacion/recetas/servicios/RecetaServiceImpl.java
@@ -1,23 +1,23 @@
 package com.ipartek.formacion.recetas.servicios;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.ipartek.formacion.recetas.dtos.Favoritos;
 import com.ipartek.formacion.recetas.entidades.Dificultad;
 import com.ipartek.formacion.recetas.entidades.Ingrediente;
 import com.ipartek.formacion.recetas.entidades.Plato;
 import com.ipartek.formacion.recetas.entidades.PlatoIngrediente;
 import com.ipartek.formacion.recetas.entidades.TipoCocina;
-import com.ipartek.formacion.recetas.entidades.Usuario;
 import com.ipartek.formacion.recetas.repositorios.DificultadRepository;
 import com.ipartek.formacion.recetas.repositorios.IngredienteRepository;
 import com.ipartek.formacion.recetas.repositorios.PlatoIngredienteRepository;
 import com.ipartek.formacion.recetas.repositorios.PlatoRepository;
 import com.ipartek.formacion.recetas.repositorios.TipoCocinaRepository;
-import com.ipartek.formacion.recetas.repositorios.UsuarioRepository;
 
+import lombok.extern.java.Log;
+
+@Log
 @Service
 public class RecetaServiceImpl implements RecetaService {
 
@@ -31,43 +31,20 @@ public class RecetaServiceImpl implements RecetaService {
 	
 	private TipoCocinaRepository tipoCocinaRepository;
 	
-	public RecetaServiceImpl(PlatoIngredienteRepository platoIngredienteRepository, UsuarioRepository usuarioRepository, PlatoRepository platoRepository, IngredienteRepository ingredienteRepository, DificultadRepository dificultadRepository, TipoCocinaRepository tipoCocinaRepository) {
-		dificultadRepository.save(Dificultad.builder().nombre("Baja").puntuacion(3).build());
-		dificultadRepository.save(Dificultad.builder().nombre("Media").puntuacion(5).build());
-		dificultadRepository.save(Dificultad.builder().nombre("Alta").puntuacion(7).build());
-		
-		tipoCocinaRepository.save(TipoCocina.builder().nombre("Americana").build());
-		tipoCocinaRepository.save(TipoCocina.builder().nombre("Española").build());
-		tipoCocinaRepository.save(TipoCocina.builder().nombre("Francesa").build());
-		tipoCocinaRepository.save(TipoCocina.builder().nombre("Italiana").build());
-		
-		var tomate = Ingrediente.builder().nombre("Tomate").build();
-		var jamon = Ingrediente.builder().nombre("Jamón").build();
-		
-		ingredienteRepository.save(Ingrediente.builder().nombre("Patata").build());
-		ingredienteRepository.save(Ingrediente.builder().nombre("Lechuga").build());
-		ingredienteRepository.save(tomate);
-		ingredienteRepository.save(jamon);
-		
-		var tipoCocina = TipoCocina.builder().id(4L).build();
-		var dificultad = Dificultad.builder().id(1L).build();
-		var plato = Plato.builder().nombre("Pizza").preparacion("Lorem ipsum dolor sit amet, consectetur adipisicing elit. Laborum dignissimos vero eligendi beatae adipisci non necessitatibus tempore consequuntur soluta asperiores ipsa eaque similique quia! Architecto consequatur a totam aut fuga!").dificultad(dificultad).tipoCocina(tipoCocina).build();
-
-		platoRepository.save(plato);
-		
-		platoIngredienteRepository.save(PlatoIngrediente.builder().plato(plato).ingrediente(tomate).medida("rodajas").cantidad(new BigDecimal(5)).build());
-		platoIngredienteRepository.save(PlatoIngrediente.builder().plato(plato).ingrediente(jamon).medida("láminas").cantidad(new BigDecimal(20)).build());
-		
-		usuarioRepository.save(Usuario.builder().nombre("Javier").email("javier@email.net").password("$2a$12$mof.u/4EIo58hR7On9DnPevyqBC7kb9FHzT.LN/BjF8xOqQVTP1NO").rol("ADMIN").build());
-		usuarioRepository.save(Usuario.builder().nombre("Pepe").email("pepe@email.net").password("$2a$12$Dij9cgV3mXDQYtOo4nvQTOLaUz3URoe7DGjhBrqGa1fEEzkNVhBgq").rol("USER").build());
-		
-		this.platoRepository = platoRepository;
+	private Favoritos favoritos;
+	
+	public RecetaServiceImpl(PlatoIngredienteRepository platoIngredienteRepository,
+			IngredienteRepository ingredienteRepository, PlatoRepository platoRepository,
+			DificultadRepository dificultadRepository, TipoCocinaRepository tipoCocinaRepository, Favoritos favoritos) {
+		super();
+		this.platoIngredienteRepository = platoIngredienteRepository;
 		this.ingredienteRepository = ingredienteRepository;
+		this.platoRepository = platoRepository;
 		this.dificultadRepository = dificultadRepository;
 		this.tipoCocinaRepository = tipoCocinaRepository;
-		this.platoIngredienteRepository = platoIngredienteRepository;
+		this.favoritos = favoritos;
 	}
-	
+
 	@Override
 	public Iterable<Dificultad> listarDificultades() {
 		return dificultadRepository.findAll();
@@ -92,6 +69,7 @@ public class RecetaServiceImpl implements RecetaService {
 	public void anadirPlato(Plato plato) {
 		plato.setId(null);
 		platoRepository.save(plato);
+		log.info("Se ha añadido el plato " + plato);
 	}
 
 	@Override
@@ -102,6 +80,11 @@ public class RecetaServiceImpl implements RecetaService {
 	@Override
 	@Transactional
 	public void borrarPlato(Long id) {
+		if(!platoRepository.existsById(id)) {
+			log.warning("Se ha intentado borrar un plato que no existe: " + id);
+			throw new ServiciosException("No existe el plato a borrar");
+		}
+		
 		platoIngredienteRepository.deleteByPlatoId(id);
 		platoRepository.deleteById(id);
 	}
@@ -118,6 +101,10 @@ public class RecetaServiceImpl implements RecetaService {
 
 	@Override
 	public Plato verPlato(Long id) {
+		if(id == null) {
+			throw new ServiciosException("El id de plato no puede ser null");
+		}
+		
 		return platoRepository.findById(id).orElse(null);
 	}
 
@@ -125,4 +112,24 @@ public class RecetaServiceImpl implements RecetaService {
 	public Iterable<PlatoIngrediente> verIngredientesPlato(Long id) {
 		return platoIngredienteRepository.findByPlatoId(id);
 	}
+
+	@Override
+	public Favoritos favoritos() {
+		return favoritos;
+	}
+
+	@Override
+	public Plato agregarFavorito(Long id) {
+		var platoOptional = platoRepository.findById(id);
+		
+		if(platoOptional.isPresent()) {
+			var plato = platoOptional.get();
+			favoritos.getPlatos().put(plato.getId(), plato);
+			return plato;
+		}
+		
+		throw new ServiciosException("No se ha encontrado el plato a agregar a favoritos");
+	}
+	
+	
 }
